Flag sensitive records according to BDBSA

This will flag records from both taxa and surveys identified as sensitive by the Biological Databases of South Australia (BDBSA). Survey matches will only be done by Survey Number (although it can be named something else, as specified in surv_nr). Likewise, taxa matches will only be done by NSXCODE (which can also be named something else, as specified in nsx_col).

flag_sens_records(recs_df, nsx_col = NULL, surv_col = NULL)

Arguments

recs_df

Dataframe of taxa records with nsx_col and surv_col.

nsx_col

Character. Name of column containing BDBSA NSXCODE to check for sensitivity.

surv_nr

Character. Name of column containing BDBSA Survey Number to check for sensitivity.

Value

recs_df with extra column(s) sens_surv and sens_taxa (logical).