Get an occurrence record set from GBIF

Deprecated. Use get_galah() with atlas = "GBIF".

get_gbif(
  ...,
  aoi = NULL,
  save_dir = NULL,
  get_new = FALSE,
  request_wait = 20,
  name = "gbif",
  data_map = NULL,
  filter_inconsistent = TRUE,
  filter_NA_date = TRUE,
  occ_char = TRUE,
  adj_spa_rel = TRUE,
  previous_key = NULL
)

Arguments

...

Other arguments passed to rgbif::occ_download()

aoi

sf defining area of interest.

save_dir

Character. File path into which to save outputs. If null results will be saved to fs::path("out", "ds", "gbif") as file gbif_raw.rds.

get_new

Logical. If FALSE will attempt to load data from previously saved results.

request_wait

Integer. Time in seconds to wait between rgbif::occ_download_meta() requests. Used by rgbif::occ_download_wait() status_ping argument.

name

Character. data_name value in envImport::data_map (or other data_map)

data_map

Dataframe or NULL. Mapping of fields to retrieve. See example envImport::data_map

filter_NA_date

Logical. Filter if is.na(eventDate).

occ_char

Logical. If true, occ_derivation will be coerced to character (to match other data sources).

adj_spa_rel

Logical. If true, an attempt will be made to check coordinateUncertaintyInMeters against: information in informationWithheld. If informationWithheld contains "Coordinate uncertainty increased to", readr::parse_number() is used to retrieve that number, which is then used to replace any value in coordinateUncertaintyInMeters; and if the column issue contains COORDINATE_UNCERTAINTY_METERS_INVALID, coordinateUncertaintyInMeters is limited to 10000 or greater.

previous_key

Character. e.g. 0092123-240506114902167. If provided, an attempt will be made to load from previously successful download of occurrence data.

Value

Dataframe, save_file, gbif_data_ref.bib (in the same directory as save_file) and full GBIF download.

Details

Get new occurrence record set from GBIF and save as .rds. GBIF is the Global Biodiversity Information Facility.

Uses various rgbif functions to return a dataframe of occurence records. Requires gbif credentials.

Any arguments to rgbif::occ_download() can be passed via .... For convenience, aoi can also be passed directly and internally it is converted to a bounding box in appropriate lat/long and passed to rgbif::pred_within() in WKT format.

Examples

if (FALSE) {
# Australian Bustard Ardeotis australis with year == 1901. Returns 7 records
# but can take a while to run depending on GBIF waiting times.
get_gbif(aoi = NULL, save_dir = NULL, get_new = FALSE, rgbif::pred("taxonKey", 2474903), rgbif::pred("year", 1901))
}