Unite (assemble) disparate data sources

Looks for any object with the names given by data_map$data_name and attempts to unite them into one data frame (tibble).

unite_data(
  df_to_unite,
  data_map,
  out_file,
  get_new = FALSE,
  add_month = TRUE,
  add_year = TRUE,
  make_occ = TRUE,
  occ_cols = c("occ_derivation", "quantity"),
  absences = c("0", "none detected", "none observed", "None detected", "ABSENT"),
  previous = c("delete", "move"),
  compare_previous = TRUE,
  compare_cols = c("data_name", "survey")
)

Arguments

df_to_unite

Dataframe with (at least) columns data_name and a list column of results from make_data_name

data_map

Dataframe or NULL. Mapping of fields to retrieve. See example envImport::data_map

out_file

Path to save results to. Will always save .parquet irrespective of any file type implied by out_file.

get_new

Logical. If FALSE, will attempt to load from existing out_file

add_month, add_year

Logical. Add a year and/or month column to returned data frame (requires a date field to be specified by data_map)

make_occ

Logical. Make an occ column (occurrence) of 1 = detected, 0 = not detected? Due to the plethora of ways original data sets record numbers and absences this should not be considered 100% reliable.

absences

Character. If make_occ what values are considered absences?

previous

Character. What to do with any previous out_file. Default is 'delete'. Alternative 'move' will rename to the same location as gsub("\.parquet", paste0("moved__", format(now(), "%Y%m%d_%H%M%S"), ".parquet"), out_file)

compare_previous

Logical. If TRUE a comparison of records per compare_cols will be made between the new and previous out_file. Ignored unless previous == "move

compare_cols

If compare_previous which columns to comapare. Default is data_name and survey.

Value

Dataframe of united objects with columns named as per the columns of data_map, optionally with month and year columns too. out_file is also created.